requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
aiobotocore
openpyxl
pandas
python-multipart
passlib[bcrypt] 
python-jose[cryptography]
jwt
pydantic[email]
asyncpg
